Rename hook files to use-*
