+++ src/lib/seam/devices/index.ts
@@ -1,1 +1,1 @@
-export * from './list.js'
+export * from './use-devices.js'
