+++ src/lib/seam/index.ts
@@ -1,2 +1,2 @@
-export * from './client.js'
 export * from './devices/index.js'
+export * from './use-seam-client.js'
